fix some issues with relative BX

+++ Analysis/interface/LhcFills.h
@@ -28,20 +28,20 @@ class LhcFills {
   std::vector<unsigned long> getFills();
 
   // get info about a particular fill
-  std::vector<unsigned long> getRuns(unsigned fill);
-  std::string getFillingScheme(unsigned fill);
-  std::vector<unsigned> getCollisions(unsigned fill);
-  std::vector<unsigned> getBunches(unsigned fill);
-  std::vector<bool> getMask(unsigned fill);
-  std::vector<bool> getLifetimeMask(unsigned fill);
+  std::vector<unsigned long> getRuns(unsigned long fill);
+  std::string getFillingScheme(unsigned long fill);
+  std::vector<unsigned long> getCollisions(unsigned long fill);
+  std::vector<unsigned long> getBunches(unsigned long fill);
+  std::vector<bool> getMask(unsigned long fill);
+  std::vector<bool> getLifetimeMask(unsigned long fill);
 
   // get info about a particular run
   unsigned long getFillFromRun(unsigned long run);
   std::string getFillingSchemeFromRun(unsigned long run);
-  std::vector<unsigned> getCollisionsFromRun(unsigned long run);
-  std::vector<unsigned> getBunchesFromRun(unsigned long run);
-  std::vector<bool> getMaskFromRun(unsigned run);
-  std::vector<bool> getLifetimeMaskFromRun(unsigned run);
+  std::vector<unsigned long> getCollisionsFromRun(unsigned long run);
+  std::vector<unsigned long> getBunchesFromRun(unsigned long run);
+  std::vector<bool> getMaskFromRun(unsigned long run);
+  std::vector<bool> getLifetimeMaskFromRun(unsigned long run);
 
   // get info about a particular filling scheme
   //   std::vector<unsigned> getCollisionsFromScheme(std::string scheme);
@@ -59,8 +59,8 @@ class LhcFills {
   bool blankString(std::string& s);
 
   // get indices
-  unsigned getIndexFromFill(unsigned fill);
-  unsigned getIndexFromRun(unsigned run);
+  unsigned long getIndexFromFill(unsigned long fill);
+  unsigned long getIndexFromRun(unsigned long run);
 
  private:
 
@@ -68,10 +68,10 @@ class LhcFills {
     unsigned long number;
     std::string scheme;
     std::vector<unsigned long> runs;
-    std::vector<unsigned> beam1;
-    std::vector<unsigned> beam2;
-    std::vector<unsigned> collisions;
-    std::vector<unsigned> bunches;
+    std::vector<unsigned long> beam1;
+    std::vector<unsigned long> beam2;
+    std::vector<unsigned long> collisions;
+    std::vector<unsigned long> bunches;
     std::vector<bool> mask;         // masked bunches (usually any bunch +/- 1 BX)
     std::vector<bool> lifetimeMask; // lifetime dependent mask
   };
@@ -83,8 +83,8 @@ class LhcFills {
 /*   std::vector< std::vector<unsigned> > bunches_; */
   
   // vector which stores index of fill for fast access
-  std::vector<unsigned> lookupFillIndex_;
-  std::vector<unsigned> lookupRunIndex_;
+  std::vector<unsigned long> lookupFillIndex_;
+  std::vector<unsigned long> lookupRunIndex_;
 
 };
